UE4.11 - D3D12 drop from changelist 2798257 - Fix for reference counting on builtin samplers when exiting
Update the PC D3D12 RHI:
- Multiple root signatures
- Command allocator refactor (cmd context uses a single cmd allocator per frame)
- Optimize setting vertex buffers
- Use CopyDescriptorsSimple() instead of CopyDescriptors() where possible
- Disable draw clears by default with a console variable (D3D12 has fast clears). There is an issue with Draw clears where it's not restoring the constant buffers correctly.

#jira UE-24967

[CL 2816571 by Rolando Caloca in 4.11 branch]

+++ Engine/Source/Runtime/Windows/D3D12RHI/Private/D3D12ConstantBuffer.cpp
@@ -1,4 +1,4 @@
-// Copyright 1998-2016 Epic Games, Inc. All Rights Reserved.
+// Copyright 1998-2014 Epic Games, Inc. All Rights Reserved.
 
 /*=============================================================================
 D3D12ConstantBuffer.cpp: D3D Constant buffer RHI implementation.
@@ -18,7 +18,7 @@ namespace D3D12RHI
 	const uint32 GConstantBufferSizes[MAX_CONSTANT_BUFFER_SLOTS] =
 	{
 		// CBs must be a multiple of 16
-		Align(MAX_GLOBAL_CONSTANT_BUFFER_SIZE, 16),
+		(uint32)Align(MAX_GLOBAL_CONSTANT_BUFFER_SIZE, 16),
 	};
 }
